feat(openarm): persistent software homing offset + fixed start pose
Make the OpenArm start position reproducible for episode collection and
inference. Zero now lives in the calibration file as a software homing
offset applied on every read (subtract) and MIT position write (add), so
the logical frame is stable across sessions/power cycles without relying
on the firmware-dependent 0xFE flash persistence. connect() no longer
re-burns the motor zero when a persistent (non-zero) offset calibration
is present (rezero_on_connect auto), and a start-position sanity gate
rejects a grossly wrong frame before any motion. The captured "ready"
pose (initial_pose_path) rides on top and covers record and inference
alike (both go through robot.connect()).

- damiao: software homing offsets (read/write transform), set_zero clears
  a stale offset with a warning, read_raw_positions for calibration, and
  a handshake fix (decode the motor RESPONSE, not the sent enable command,
  which used to seed the cache with a +position_max garbage position).
- openarm_follower/leader (+ bimanual): rezero_on_connect and
  start_position_tolerance_deg config, gated connect-time zero, calibrate()
  now measures the homing offset instead of burning 0xFE.
- motors_bus: MotorCalibration.homing_offset int -> float (Damiao needs
  degrees; SerialMotorsBus.write already int-casts so Feetech/Dynamixel
  are unaffected).

Backward compatible: legacy calibrations (offset 0) keep the old
re-zero-every-connect behavior. Gripper pinned to its closed zero.
Hardware validation on Thor still pending.

--- 平易な説明（レビュー用 / Kota 向け）---
腕の角度の数え方（零点）を一度決めたら固定し、テレオペ/収集/推論を毎回
同じ構えから始められるようにした変更。
- 従来: 起動のたびにモータへ「今が0」と焼き付け → 毎回真下に構え直さないと
  leader/follower の零がずれて teleop で暴れた。
- 変更: モータには焼かず、lerobot が較正ファイルに補正メモ(homing offset)を
  持ち、読みで引き算・書きで足し算。生角度は絶対エンコーダで電源断でも再現
  するので零が永続する（SO-101 / PR #655 と同じ発想の Damiao 版）。
- その上で開始"構え"を手で1回キャプチャ(capture_initial_pose)し、起動時に
  両腕がそこへ自動整列。record も inference も同じ connect 経路なので両対応。
- 後方互換: 旧較正(offset 0)は従来挙動のまま自動フォールバック。グリッパは
  閉じた零にピン留め。起動時サニティゲートで零フレーム破損を動く前に停止。
設計文書: orboh-company-vault 07.../Kento/01-design/openarm-persistent-zero-start-pose.md
実機(Thor)検証は未。

+++ src/lerobot/motors/motors_bus.py
@@ -175,7 +175,11 @@ class MotorNormMode(str, Enum):
 class MotorCalibration:
     id: int
     drive_mode: int
-    homing_offset: int
+    # Serial buses (Dynamixel/Feetech) store integer encoder ticks; the Damiao CAN
+    # bus stores a float software offset in DEGREES (motors can't persist a
+    # calibration internally, see DamiaoMotorsBus). `float` accepts both, and
+    # draccus loads legacy int JSON values into a float field without complaint.
+    homing_offset: float
     range_min: int
     range_max: int
